append data to the database file if it already exists

+++ src/storage/storage.rs
@@ -1,4 +1,4 @@
-use std::fs::File;
+use std::fs::{File, OpenOptions};
 use std::io::Write;
 use std::path::Path;
 use std::io::Result;
@@ -17,11 +17,29 @@ pub fn put(data_map: &mut HashMap<String, String>, key: &String, value: &String)
 
 fn save_to_disk(data_map: &mut HashMap<String, String>) -> Result<()> {
 
+	let exists: bool;
 	let file_path = Path::new("./db_file").join("db.txt");
-	let mut file = File::create(file_path).expect("Erro");
-	let encoded: Vec<u8> = bincode::serialize(&data_map).unwrap();
-	file.write_all(&encoded)?;
-	Ok(())
+	exists = Path::new("./db_file/db.txt").exists();
+
+	if exists == true {
+		
+		let mut file = OpenOptions::new()
+					.append(true)
+					.open(file_path)
+					.expect("cannot open this file");
+		
+		let encoded: Vec<u8> = bincode::serialize(&data_map).unwrap();
+		file.write_all(&encoded)?;
+		Ok(())
+	}
+	else {
+
+		let mut file = File::create(file_path).expect("Erro");
+		let encoded: Vec<u8> = bincode::serialize(&data_map).unwrap();
+		file.write_all(&encoded)?;
+		Ok(())
+	}
+
 }
